split lnd into lndbtc, lndltc, lndbtc-simnet and lndltc-simnet
