--- images/lnd/.dockerignore
@@ -1,1 +1,0 @@
-.src/**/.git
