+++ images/lndbtc-simnet/.dockerignore
@@ -1,0 +1,1 @@
+.src/.git
