+++ images/lndbtc-simnet/Dockerfile
@@ -1,12 +1,10 @@
 FROM golang:1.13-alpine3.11 as builder
 RUN apk add --no-cache bash git make gcc musl-dev
 WORKDIR $GOPATH/src/github.com/lightningnetwork/lnd
-ARG SRC_DIR
-ADD $SRC_DIR .
+ADD .src .
 RUN go mod vendor
 # patching
-ARG PATCHES_DIR
-ADD $PATCHES_DIR /patches/
+ADD patches /patches/
 RUN /patches/apply.sh
 # build
 ARG TAGS
@@ -18,7 +16,6 @@ RUN strip /go/bin/lnd /go/bin/lncli
 FROM alpine:3.11
 RUN apk add --no-cache bash tor
 COPY --from=builder /go/bin/lnd /go/bin/lncli /usr/local/bin/
-ARG ENTRYPOINT_FILE
-COPY $ENTRYPOINT_FILE /entrypoint.sh
+COPY entrypoint.sh /entrypoint.sh
 ENTRYPOINT ["/entrypoint.sh"]
 EXPOSE 10009
