+++ images/lndbtc-simnet/entrypoint.sh
