+++ images/lndbtc/.dockerignore
@@ -1,0 +1,1 @@
+.src/.git
