+++ images/lndbtc/entrypoint.sh
