+++ images/lndbtc/lnd.conf
