+++ images/lndbtc/wait-file.sh
