+++ images/lndltc/Dockerfile
@@ -1,19 +1,17 @@
 FROM golang:1.13-alpine3.11 as builder
 RUN apk add --no-cache bash git make gcc musl-dev
 WORKDIR $GOPATH/src/github.com/lightningnetwork/lnd
-ARG SRC_DIR
-ADD $SRC_DIR .
-RUN make tags="autopilotrpc chainrpc experimental invoicesrpc routerrpc signrpc walletrpc watchtowerrpc wtclientrpc" install
+ADD .src .
+RUN go install -v -tags="$TAGS" -ldflags "$LDFLAGS" ./cmd/lnd ./cmd/lncli
+RUN strip /go/bin/lnd /go/bin/lncli
 
 # Final stage
 FROM alpine:3.11
 RUN apk add --no-cache bash expect supervisor tor
 COPY --from=builder /go/bin/lnd /go/bin/lncli /usr/local/bin/
-ARG ENTRYPOINT_FILE
-COPY $ENTRYPOINT_FILE /entrypoint.sh
+COPY entrypoint.sh /entrypoint.sh
 COPY wait-file.sh start_tor.sh /
-ARG LND_CONF_FILE
-COPY $LND_CONF_FILE /root/lnd.conf
+COPY lnd.conf /root/lnd.conf
 COPY supervisord.conf /etc/supervisor/conf.d/
 COPY torrc /etc/tor/
 RUN mkdir -p /root/.lnd/tor
