+++ images/lndltc/lnd.conf
