+++ images/lndltc/supervisord.conf
@@ -1,0 +1,15 @@
+[supervisord]
+nodaemon=true
+logfile=/root/supervisord.log
+childlogdir=/root
+user=root
+
+[program:tor]
+stdout_logfile=/dev/stdout
+stdout_logfile_maxbytes=0
+command=/start_tor.sh
+
+[program:lnd]
+stdout_logfile=/dev/stdout
+stdout_logfile_maxbytes=0
+command=/entrypoint.sh
